refactor: config naming
