+++ packages/blocks/src/_common/widgets/format-bar/components/inline-items.ts
@@ -1,7 +1,7 @@
 import { html } from 'lit';
 
-import { formatConfigs } from '../../../configs/format/format.js';
-import { isFormatSupported } from '../../../configs/format/utils.js';
+import { textFormatConfigs } from '../../../configs/text-format/config.js';
+import { isFormatSupported } from '../../../configs/text-format/utils.js';
 import type { AffineFormatBarWidget } from '../format-bar.js';
 import { BackgroundButton } from './background/background-button.js';
 
@@ -14,7 +14,7 @@ export const InlineItems = (formatBar: AffineFormatBarWidget) => {
 
   const backgroundButton = BackgroundButton(formatBar);
 
-  return html`${formatConfigs.map(
+  return html`${textFormatConfigs.map(
       ({ id, name, icon, action, activeWhen }) =>
         html`<icon-button
           size="32px"
